Refactor common code to abstract HAL class
