+++ src/main/java/oshi/hardware/platform/linux/LinuxHardwareAbstractionLayer.java
@@ -16,39 +16,20 @@
  */
 package oshi.hardware.platform.linux;
 
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonBuilderFactory;
-import javax.json.JsonObject;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import oshi.hardware.CentralProcessor;
 import oshi.hardware.Display;
 import oshi.hardware.GlobalMemory;
-import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.PowerSource;
 import oshi.hardware.Sensors;
-import oshi.json.NullAwareJsonObjectBuilder;
+import oshi.hardware.common.AbstractHardwareAbstractionLayer;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.linux.LinuxFileSystem;
 
-/**
- * @author alessandro[at]perucchi[dot]org
- */
+public class LinuxHardwareAbstractionLayer extends AbstractHardwareAbstractionLayer {
 
-public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
-    private static final Logger LOG = LoggerFactory.getLogger(LinuxHardwareAbstractionLayer.class);
-
-    private CentralProcessor processor;
-
-    private GlobalMemory memory;
-
-    private Sensors sensors;
-
-    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public GlobalMemory getMemory() {
         if (this.memory == null) {
@@ -57,6 +38,9 @@ public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
         return this.memory;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public CentralProcessor getProcessor() {
         if (this.processor == null) {
@@ -65,46 +49,38 @@ public class LinuxHardwareAbstractionLayer implements HardwareAbstractionLayer {
         return this.processor;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public PowerSource[] getPowerSources() {
         return LinuxPowerSource.getPowerSources();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public OSFileStore[] getFileStores() {
         return LinuxFileSystem.getFileStores();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Display[] getDisplays() {
         return LinuxDisplay.getDisplays();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Sensors getSensors() {
         if (this.sensors == null) {
             this.sensors = new LinuxSensors();
         }
         return this.sensors;
-    }
-
-    @Override
-    public JsonObject toJSON() {
-        JsonArrayBuilder powerSourceArrayBuilder = jsonFactory.createArrayBuilder();
-        for (PowerSource powerSource : getPowerSources()) {
-            powerSourceArrayBuilder.add(powerSource.toJSON());
-        }
-        JsonArrayBuilder fileStoreArrayBuilder = jsonFactory.createArrayBuilder();
-        for (OSFileStore fileStore : getFileStores()) {
-            fileStoreArrayBuilder.add(fileStore.toJSON());
-        }
-        JsonArrayBuilder displayArrayBuilder = jsonFactory.createArrayBuilder();
-        for (Display display : getDisplays()) {
-            displayArrayBuilder.add(display.toJSON());
-        }
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder())
-                .add("processor", getProcessor().toJSON()).add("memory", getMemory().toJSON())
-                .add("powerSources", powerSourceArrayBuilder.build()).add("fileStores", fileStoreArrayBuilder.build())
-                .add("displays", displayArrayBuilder.build()).add("sensors", getSensors().toJSON()).build();
     }
 }
