+++ src/main/java/oshi/hardware/platform/mac/MacHardwareAbstractionLayer.java
@@ -16,41 +16,19 @@
  */
 package oshi.hardware.platform.mac;
 
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonBuilderFactory;
-import javax.json.JsonObject;
-
 import oshi.hardware.CentralProcessor;
 import oshi.hardware.Display;
 import oshi.hardware.GlobalMemory;
-import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.PowerSource;
 import oshi.hardware.Sensors;
-import oshi.json.NullAwareJsonObjectBuilder;
+import oshi.hardware.common.AbstractHardwareAbstractionLayer;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.mac.MacFileSystem;
 
-/**
- * @author alessandro[at]perucchi[dot]org
- * @author widdis[at]gmail[dot]com
- */
-public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
+public class MacHardwareAbstractionLayer extends AbstractHardwareAbstractionLayer {
 
-    private CentralProcessor processor;
-
-    private GlobalMemory _memory;
-
-    private PowerSource[] _powerSources;
-
-    private Sensors sensors;
-
-    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see oshi.hardware.HardwareAbstractionLayer#getProcessor()
+    /**
+     * {@inheritDoc}
      */
     @Override
     public CentralProcessor getProcessor() {
@@ -60,62 +38,49 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
         return this.processor;
     }
 
-    /*
-     * (non-Javadoc)
-     * 
-     * @see oshi.hardware.HardwareAbstractionLayer#getMemory()
+    /**
+     * {@inheritDoc}
      */
     @Override
     public GlobalMemory getMemory() {
-        if (this._memory == null) {
-            this._memory = new MacGlobalMemory();
+        if (this.memory == null) {
+            this.memory = new MacGlobalMemory();
         }
-        return this._memory;
+        return this.memory;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public PowerSource[] getPowerSources() {
-        if (this._powerSources == null) {
-            this._powerSources = MacPowerSource.getPowerSources();
-        }
-        return this._powerSources;
+        return MacPowerSource.getPowerSources();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public OSFileStore[] getFileStores() {
         return MacFileSystem.getFileStores();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Display[] getDisplays() {
         return MacDisplay.getDisplays();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Sensors getSensors() {
         if (this.sensors == null) {
             this.sensors = new MacSensors();
         }
         return this.sensors;
-    }
-
-    @Override
-    public JsonObject toJSON() {
-        JsonArrayBuilder powerSourceArrayBuilder = jsonFactory.createArrayBuilder();
-        for (PowerSource powerSource : getPowerSources()) {
-            powerSourceArrayBuilder.add(powerSource.toJSON());
-        }
-        JsonArrayBuilder fileStoreArrayBuilder = jsonFactory.createArrayBuilder();
-        for (OSFileStore fileStore : getFileStores()) {
-            fileStoreArrayBuilder.add(fileStore.toJSON());
-        }
-        JsonArrayBuilder displayArrayBuilder = jsonFactory.createArrayBuilder();
-        for (Display display : getDisplays()) {
-            displayArrayBuilder.add(display.toJSON());
-        }
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder()).add("processor", getProcessor().toJSON())
-                .add("memory", getMemory().toJSON()).add("powerSources", powerSourceArrayBuilder.build())
-                .add("fileStores", fileStoreArrayBuilder.build()).add("displays", displayArrayBuilder.build())
-                .add("sensors", getSensors().toJSON()).build();
     }
 }
